Updated README.md

+++ KHRay/KHRay/Vector3.h
@@ -149,8 +149,8 @@ template <typename T>
 template <typename T>
 [[nodiscard]] inline Vector3<T> Cross(const Vector3<T>& v1, const Vector3<T>& v2)
 {
-	double v1x = v1.x, v1y = v1.y, v1z = v1.z;
-	double v2x = v2.x, v2y = v2.y, v2z = v2.z;
+	float v1x = v1.x, v1y = v1.y, v1z = v1.z;
+	float v2x = v2.x, v2y = v2.y, v2z = v2.z;
 	return Vector3<T>((v1y * v2z) - (v1z * v2y),
 		(v1z * v2x) - (v1x * v2z),
 		(v1x * v2y) - (v1y * v2x));
@@ -172,11 +172,11 @@ inline void CoordinateSystem(const Vector3<T>& v1, Vector3<T>* v2, Vector3<T>* v
 {
 	if (std::abs(v1.x) > std::abs(v1.y))
 	{
-		*v2 = Vector3<T>(-v1.z, 0, v1.x) / std::sqrt(v1.x * v1.x + v1.z * v1.z);
+		*v2 = Vector3<T>(-v1.z, 0.0f, v1.x) / std::sqrt(v1.x * v1.x + v1.z * v1.z);
 	}
 	else
 	{
-		*v2 = Vector3<T>(0, v1.z, -v1.y) / std::sqrt(v1.y * v1.y + v1.z * v1.z);
+		*v2 = Vector3<T>(0.0f, v1.z, -v1.y) / std::sqrt(v1.y * v1.y + v1.z * v1.z);
 	}
 	*v3 = Normalize(Cross(v1, *v2));
 }
